+++ flight_sim/.gitignore
@@ -1,0 +1,3 @@
+/build/
+
+.cache
