+++ flight_sim/README.md
@@ -1,3 +1,7 @@
 # Flight Simulator
 
 Flight simulation branch.
+
+## Build instructions
+
+
