+++ flight_sim/src/main.cpp
@@ -1,0 +1,10 @@
+/*
+ * This is the CMAKE target file, delete later on after
+ * changing target rules
+ */
+
+#include <stdio.h>
+
+int main() {
+    printf("Hello\n");
+}
